Move logger script testing to other file
pipeline fails when import of other test case was executed before
