+++ test/test_scripts.py
@@ -7,8 +7,6 @@ This module tests that the scripts are all callable.
 
 import subprocess
 import unittest
-from unittest import mock
-from unittest.mock import Mock
 import sys
 import errno
 from abc import ABCMeta, abstractmethod
@@ -87,38 +85,6 @@ class TestLoggerScript(CanScriptTest):
         import can.logger as module
 
         return module
-
-    def test_log_virtual(self):
-        import can
-
-        msg = can.Message(
-            arbitration_id=0xC0FFEE, data=[0, 25, 0, 1, 3, 1, 4, 1], is_extended_id=True
-        )
-
-        # Patch VirtualBus object
-        patcher_virtual_bus = mock.patch("can.interfaces.virtual.VirtualBus", spec=True)
-        self.MockVirtualBus = patcher_virtual_bus.start()
-        self.addCleanup(patcher_virtual_bus.stop)
-        mock_virtual_bus = self.MockVirtualBus.return_value
-        mock_virtual_bus.recv = Mock(side_effect=[msg, KeyboardInterrupt])
-        mock_virtual_bus.shutdown = Mock()
-
-        # Patch Logger object
-        patcher_logger = mock.patch("can.Logger", spec=True)
-        self.MockLogger = patcher_logger.start()
-        self.addCleanup(patcher_logger.stop)
-        mock_logger = self.MockLogger.return_value
-        mock_logger.stop = Mock()
-
-        module = self._import()
-
-        sys.argv = [sys.argv[0], "-i", "virtual"]
-        module.main()
-        self.MockVirtualBus.assert_called_once()
-        mock_virtual_bus.shutdown.assert_called_once()
-
-        self.MockLogger.assert_called_once()
-        mock_logger.stop.assert_called_once()
 
 
 class TestPlayerScript(CanScriptTest):
